+++ shared/amdgpu-windows-interop/pal/inc/util/palMutex.h
@@ -1,7 +1,7 @@
 /*
  ***********************************************************************************************************************
  *
- *  Copyright (c) 2014-2025 Advanced Micro Devices, Inc. All Rights Reserved.
+ *  Copyright (c) Advanced Micro Devices, Inc., or its affiliates. All rights reserved.
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy
  *  of this software and associated documentation files (the "Software"), to deal
@@ -145,7 +145,7 @@ public:
     /// Defines RWLockData as a Windows RWLOCK
     typedef SRWLOCK RWLockData;
     RWLock() noexcept : m_osRWLock {} { InitializeSRWLock(&m_osRWLock); }
-    ~RWLock() noexcept { /* No Win32 destory function */ };
+    ~RWLock() noexcept { /* No Win32 destroy function */ };
 #else
     /// Defines RWLockData as a unix pthread_rwlock_t
     typedef pthread_rwlock_t  RWLockData;
@@ -297,7 +297,7 @@ extern uint32 AtomicDecrement(volatile uint32* pValue);
 /// @param [in,out] pValue Pointer to the value to be decremented.
 ///
 /// @returns Result of the decrement operation.
-extern uint32 AtomicDecrement64(volatile uint64* pValue);
+extern uint64 AtomicDecrement64(volatile uint64* pValue);
 
 /// Performs an atomic compare and swap operation on two 32-bit unsigned integers. This operation compares *pTarget
 /// with oldValue and replaces it with newValue if they match. If the values don't match, no action is taken.
